// src/types.ts
export interface SegmentEffort {
  id: number;
  name: string;
  elapsed_time: number;
  moving_time: number;
  start_date: string;
  start_date_local: string;
  distance: number;
  average_watts?: number;
  average_heartrate?: number;
  max_heartrate?: number;
  average_cadence?: number;
  pr_rank: number | null;
  segment: {
    id: number;
    name: string;
    distance: number;
    average_grade: number;
    maximum_grade: number;
    elevation_high: number;
    elevation_low: number;
    climb_category: number;
    city: string | null;
    state: string | null;
    country: string | null;
  };
}
// ...
export interface SegmentDetail {
  id: number;
  name: string;
  activity_type: string;
  distance: number;
  average_grade: number;
  maximum_grade: number;
  elevation_high: number;
  elevation_low: number;
  start_latlng: [number, number];
  end_latlng: [number, number];
  climb_category: number;
  city: string | null;
  state: string | null;
  country: string | null;
  total_elevation_gain: number;
  effort_count: number;
  athlete_count: number;
  hazardous: boolean;
  star_count: number;
  created_at: string;
  updated_at: string;
  map?: {
    id: string;
    polyline: string;
  };
  athlete_pr_effort?: SegmentEffort;
  xoms?: {
    kom: string;
    qom: string;
    overall: string;
  };
}

/** Segment explorer result */
export interface ExplorerSegment {
  id: number;
  name: string;
  climb_category: number;
  climb_category_desc: string;
  avg_grade: number;
  start_latlng: [number, number];
  end_latlng: [number, number];
  elev_difference: number;
  distance: number;
  points: string;
}